+++ src/Error Control System/ErrorControlSystem/Shared/UI/ExceptionMapper/TreeMapItemMode.cs
@@ -1,0 +1,10 @@
+﻿
+
+namespace ErrorControlSystem.Shared.UI.ExceptionMapper
+{
+    public enum TreeMapItemMode
+    {
+        Exception,
+        InnerException
+    }
+}