+++ src/Tests/ErrorControlSystem.Shared.UI.ExceptionMapper.Tests/TestClassB.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace ErrorControlSystem.Shared.UI.ExceptionMapper.Tests
+{
+    public static class TestClassB
+    {
+        public static void B1()
+        {
+            try
+            {
+                TestClassC.C1();
+            }
+            catch (Exception exp)
+            {
+                throw new AggregateException(exp);
+            }
+        }
+    }
+}