+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Added
 * Bug fixes
 * Add Server Instance Username and Password Text boxes to Analyzer #40
 * Solve Analyzer Throw an Exception to Can Not Found Database #38 issue
+* Developer UI designed by WPF user controls to show full info message